+++ src/main/java/carrace/moving/Moving.java
@@ -1,5 +1,5 @@
 package carrace.moving;
 
 public interface Moving {
-    boolean isMove();
+    int move();
 }
