+++ src/main/java/carrace/moving/RandomMoving.java
@@ -8,13 +8,8 @@ public class RandomMoving implements Moving {
     private static final int STOP = 0;
 
     private final RandomNumberGenerator generator = new RandomNumberGenerator();
-
     @Override
-    public boolean isMove() {
-        return generator.getRandomBetweenZeroAndNine() >= DECISION_VALUE;
-    }
-
-    public int move(boolean isMove) {
-        return isMove ? GO : STOP;
+    public int move() {
+        return generator.getRandomBetweenZeroAndNine() >= DECISION_VALUE ? GO : STOP;
     }
 }
